rename addshortinsulin static lib

+++ Diabetto.iOS.Intents/Diabetto.iOS.Intents.AddShortInsulinStaticLib/bo/StructsAndEnums.cs
@@ -1,11 +1,10 @@
-using System;
 using ObjCRuntime;
 
 namespace Diabetto
 {
 	[Watch (5,0), NoTV, NoMac, iOS (12,0)]
 	[Native]
-	public enum AddShortInsulinIntentResponseCode : nint
+	public enum AddShortInsulinIntentResponseCode : long
 	{
 		Unspecified = 0,
 		Ready,
@@ -18,7 +17,7 @@ namespace Diabetto
 
 	[Watch (6,0), iOS (13,0)]
 	[Native]
-	public enum AddShortInsulinShortInsulinUnsupportedReason : nint
+	public enum AddShortInsulinShortInsulinUnsupportedReason : long
 	{
 		NegativeNumbersNotSupported = 1,
 		GreaterThanMaximumValue,
@@ -27,7 +26,7 @@ namespace Diabetto
 
 	[Watch (6,0), iOS (13,0)]
 	[Native]
-	public enum AddShortInsulinBreadUnitsUnsupportedReason : nint
+	public enum AddShortInsulinBreadUnitsUnsupportedReason : long
 	{
 		NegativeNumbersNotSupported = 1,
 		GreaterThanMaximumValue,
